pyproject.toml
[project]
name = "multi-pdf-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "huggingface-hub==0.32.3",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-huggingface>=0.2.0",
    "openai>=1.82.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "tiktoken>=0.9.0",
]
